fixed wrong order of config initialization
